Increase size for node 'id' field to prevent overflow.
For RTL88**A, RTWN_MACID_VALID | R12A_MACID_MAX == RTWN_MACID_UNDEFINED;
(i.e., there was no way to encode it properly). To prevent such situations,
the 'macid' variable size was increazed (uint8_t -> int).

+++ sys/dev/rtwn/if_rtwn.c
@@ -1680,7 +1680,7 @@ rtwn_newassoc(struct ieee80211_node *ni, int isnew)
 {
 	struct rtwn_softc *sc = ni->ni_ic->ic_softc;
 	struct rtwn_node *un = RTWN_NODE(ni);
-	uint8_t id;
+	int id;
 
 	if (!isnew)
 		return;
