+++ sys/dev/rtwn/rtl8188e/r88e.h
@@ -58,7 +58,7 @@ int	r88e_fw_cmd(struct rtwn_softc *, uint8_t, const void *, int);
 void	r88e_fw_reset(struct rtwn_softc *, int);
 #endif
 void	r88e_macid_enable_link(struct rtwn_softc *, int, int);
-void	r88e_set_media_status(struct rtwn_softc *, uint8_t);
+void	r88e_set_media_status(struct rtwn_softc *, int);
 #ifndef RTWN_WITHOUT_UCODE
 int	r88e_set_rsvd_page(struct rtwn_softc *, int, int, int);
 int	r88e_set_pwrmode(struct rtwn_softc *, struct ieee80211vap *, int);
